Remove classic readiness check
