+++ quarkus/config-api/src/main/java/org/keycloak/config/HealthOptions.java
@@ -9,12 +9,4 @@ public class HealthOptions {
             .buildTime(true)
             .build();
 
-    public static final Option<Boolean> HEALTH_CLASSIC_PROBES_ENABLED = new OptionBuilder<>("health-classic-probes-enabled", Boolean.class)
-            .category(OptionCategory.HEALTH)
-            .description("If enabled, use the original Quarkus blocking handlers for '/health/ready' and '/health/live' endpoints.")
-            .defaultValue(Boolean.FALSE)
-            .buildTime(true)
-            .hidden()
-            .build();
-
 }
